lab 3/requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
mlflow==2.11.2
scikit-learn==1.4.0
pandas==2.2.1